% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_weekcal.R
\name{build_weekcal}
\alias{build_weekcal}
\title{Build Week calendar}
\usage{
build_weekcal(start = 2018, end = 2020)
}
\arguments{
\item{start}{Starting year}

\item{end}{Ending year}
}
\value{
a \code{data.table}
}
\description{
Build Week calendar
}
\examples{

build_weekcal()

}
