% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kd_coefs.R
\name{compute_kd_coefs}
\alias{compute_kd_coefs}
\title{Compute Kd coefficients}
\usage{
compute_kd_coefs(kipr, kivt, kd_cho)
}
\arguments{
\item{kipr}{Kipr coefficient from \code{\link{compute_kipr}}.}

\item{kivt}{Kivt coefficient from \code{\link{compute_kivt}}.}

\item{kd_cho}{Kd coefficients read with \code{\link{read_kd_cho}}.}
}
\value{
a \code{data.table}
}
\description{
Compute Kd coefficients
}
