% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kipr.R
\name{compute_kipr}
\alias{compute_kipr}
\title{Compute Kipr coefficient}
\usage{
compute_kipr(calendar, clusters_desc, years = NULL)
}
\arguments{
\item{calendar}{Calendar data read with \code{\link{read_calendar}}.}

\item{clusters_desc}{Clusters / groups description read with \code{\link{read_cluster_desc}}.}

\item{years}{Years to consider, if \code{NULL} (default), year range from \code{calendar} will be used.}
}
\value{
a \code{data.table}
}
\description{
Compute Kipr coefficient
}
\examples{
\dontrun{
# Calendar data
calendar <- read_calendar(path = "REF_Planning_5_ans_mars_2018.xlsx")

# Clusters description
clusters <- read_cluster_desc("HypothesesRTE_CHO-4145.xlsx")


# Kipr computation
kipr <- compute_kipr(calendar, clusters)
kipr
}
}
