% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clusters_edf.R
\name{create_clusters_edf}
\alias{create_clusters_edf}
\title{Create EDF clusters}
\usage{
create_clusters_edf(planning, opts = simOptions())
}
\arguments{
\item{planning}{Calendar data read with \code{\link{read_calendar}}.}

\item{opts}{List of simulation parameters returned by the function
\code{setSimulationPath}}
}
\description{
Create EDF clusters
}
