% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clusters_nuclear.R
\name{create_clusters_nuclear}
\alias{create_clusters_nuclear}
\title{Create nuclear clusters}
\usage{
create_clusters_nuclear(
  calendar,
  clusters_desc,
  kd_cho,
  start_date = NULL,
  area_name = NULL,
  law_planned = "geometric",
  volatility_planned = 1,
  constraints = NULL,
  opts = simOptions(),
  correspondance_filiere_cluster = NULL,
  opts_bp = NULL
)
}
\arguments{
\item{calendar}{Calendar data read with \code{\link{read_calendar}}.}

\item{clusters_desc}{Clusters / groups description read with \code{\link{read_cluster_desc}}.}

\item{kd_cho}{Kd coefficients read with \code{\link{read_kd_cho}}.}

\item{start_date}{Starting date of the study, if \code{NULL} (default),
the date will be retrieve from the Antares study.}

\item{area_name}{Name of the area where to create clusters.}

\item{law_planned}{Law to use in Antares.}

\item{volatility_planned}{Volatility for the law.}

\item{constraints}{Stretch/Zircaloy constraints read with
\code{\link{read_constraints}}. Defaults to NULL.}

\item{opts}{List of simulation parameters returned by the function
\code{setSimulationPath}}

\item{correspondance_filiere_cluster}{correspondance_filiere_cluster read from correspondance_filiere_cluster.xlsx.}

\item{opts_bp}{List of simulation parameters returned by the function reference study (generaly bp study)
\code{setSimulationPath}}
}
\description{
Create nuclear clusters
}
