% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clusters_other.R
\name{create_clusters_other}
\alias{create_clusters_other}
\title{Create other clusters}
\usage{
create_clusters_other(planning, infos, opts = simOptions())
}
\arguments{
\item{planning}{Calendar data read with \code{\link{read_calendar}}.}

\item{infos}{Info}

\item{opts}{List of simulation parameters returned by the function
\code{setSimulationPath}}
}
\description{
Create other clusters
}
