% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_scenarii.R
\name{mean_scenarii}
\alias{mean_scenarii}
\title{Mean of scenarii by weeks}
\usage{
mean_scenarii(first_weekday = 1, opts = simOptions())
}
\arguments{
\item{first_weekday}{The first day to use for starting a week, default to \code{1} (monday).}

\item{opts}{List of simulation parameters returned by the function
\code{setSimulationPath}}
}
\value{
a \code{data.table}.
}
\description{
Mean of scenarii by weeks
}
\examples{
\dontrun{

library(antaresRead)
setSimulationPath(path = "PATH/TO/STUDY")

}
}
