% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_thermal_ts.R
\name{mean_thermal_ts2}
\alias{mean_thermal_ts2}
\title{Mean of thermal timeseries by weeks}
\usage{
mean_thermal_ts2(first_weekday = 1, opts = simOptions())
}
\arguments{
\item{first_weekday}{The first day to use for starting a week, default to \code{1} (monday).}

\item{opts}{List of simulation parameters returned by the function
\code{setSimulationPath}}
}
\value{
a \code{data.table}.
}
\description{
Mean of thermal timeseries by weeks
}
