% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msp_run_clusters.R
\name{msp_run_clusters}
\alias{msp_run_clusters}
\title{Get sum of Min Stable Power for running clusters by weeks}
\usage{
msp_run_clusters(first_weekday = 1, area = "fr",
  remove_clusters = NULL, start_date = NULL, opts = simOptions())
}
\arguments{
\item{first_weekday}{The first day to use for starting a week, default to \code{1} (monday).}

\item{area}{Area(s) to include in the computation.}

\item{remove_clusters}{Character vector. Cluster(s) to remove.}

\item{start_date}{Starting date of the study. If \code{NULL} (the default),
the date of the antares study is used.}

\item{opts}{List of simulation parameters returned by the function
\code{setSimulationPath}}
}
\value{
a \code{data.table}.
}
\description{
Get sum of Min Stable Power for running clusters by weeks
}
\examples{
\dontrun{

library(antaresRead)
setSimulationPath(path = "PATH/TO/STUDY")

msp_run_clusters()

msp_run_clusters(first_weekday = 6) # saturday

}
}
