% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn_stop_clusters.R
\name{pcn_stop_clusters}
\alias{pcn_stop_clusters}
\title{Get sum of PCN for stopped clusters by weeks}
\usage{
pcn_stop_clusters(start_date = NULL, first_weekday = 1,
  stop_all_week = TRUE, opts = simOptions())
}
\arguments{
\item{start_date}{Starting date of the study.}

\item{first_weekday}{The first day to use for starting a week, default to \code{1} (monday).}

\item{stop_all_week}{Consider that a cluster is stopped if it is really stopped the 7 days
of the weeks, otherwise the PCN of a clusters is counted for a week if it is stopped 
at least one day in the week.}

\item{opts}{List of simulation parameters returned by the function
\code{setSimulationPath}}
}
\value{
a \code{data.table}.
}
\description{
Get sum of PCN for stopped clusters by weeks
}
\examples{
\dontrun{

library(antaresRead)
setSimulationPath(path = "PATH/TO/STUDY")

pcn_stop_clusters()

pcn_stop_clusters(first_weekday = 6) # saturday

}
}
