% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_calendar.R
\name{read_calendar}
\alias{read_calendar}
\title{Read calendar data}
\usage{
read_calendar(path)
}
\arguments{
\item{path}{Path to a file.}
}
\value{
a \code{data.table}
}
\description{
Read calendar data
}
