% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cluster_desc.R
\name{read_cluster_desc}
\alias{read_cluster_desc}
\title{Read Cluster Description}
\usage{
read_cluster_desc(path)
}
\arguments{
\item{path}{Path to Excel file}
}
\value{
a \code{data.table}
}
\description{
Read Cluster Description
}
\examples{
\dontrun{

cluster_desc <- read_cluster_desc("HypothesesRTE_CHO-4145.xlsx")

}

}
