% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_constraints.R
\name{read_constraints}
\alias{read_constraints}
\title{Read constraints}
\usage{
read_constraints(path)
}
\arguments{
\item{path}{Path to Excel file.}
}
\value{
a \code{data.table}
}
\description{
Read constraints
}
