% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_info.R
\name{read_info}
\alias{read_info}
\title{Read other clusters informations}
\usage{
read_info(path)
}
\arguments{
\item{path}{Path to file.}
}
\value{
a \code{data.table}
}
\description{
Read other clusters informations
}
