% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kd_edf.R
\name{read_kd_edf}
\alias{read_kd_edf}
\title{Read Kd coefficents for EDF clusters}
\usage{
read_kd_edf(path)
}
\arguments{
\item{path}{Path to Excel file.}
}
\value{
a \code{data.table}
}
\description{
Read Kd coefficents for EDF clusters
}
