% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_planning.R
\name{read_planning}
\alias{read_planning}
\title{Read a maintenance planning}
\usage{
read_planning(path, sheet = "Planning", start_col = "A", start_row = 16)
}
\arguments{
\item{path}{Character. Path to the Excel file}

\item{sheet}{Character. Sheet in which the planning can be found. Defaults to 'Planning'.}

\item{start_col}{Character. Leftmost column of the planning (upper-case letter). Defaults to 'A'.}

\item{start_row}{Integer. Row number of the planning's header. Defaults to 16.}
}
\value{
A 12-column \code{data.table}: the same 12 columns as in the Excel sheet. Useless rows
  are dropped, and values from the first 4 columns are repeated in order to obtain a 'tidy'
  dataset.
}
\description{
Read a maintenance planning
}
\examples{
\dontrun{

mydt <- read_planning("1805_Planning Mensuel GDF SUEZ_REF.xlsx")

}

}
