% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_planning_rte.R
\name{read_planning_rte}
\alias{read_planning_rte}
\title{Read Rte planning}
\usage{
read_planning_rte(path, clusters_desc)
}
\arguments{
\item{path}{Path to file.}

\item{clusters_desc}{Clusters description.}
}
\value{
a \code{data.table}
}
\description{
Read Rte planning
}
