% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn_stop_clusters.R
\name{pcn_stop_clusters}
\alias{pcn_stop_clusters}
\title{Get sum of PCN for stopped clusters by weeks}
\usage{
pcn_stop_clusters(
  first_weekday = 1,
  area = "fr",
  remove_clusters = NULL,
  start_date = NULL,
  opts = simOptions()
)
}
\arguments{
\item{first_weekday}{The first day to use for starting a week, default to \code{1} (monday).}

\item{area}{Area(s) to include in the computation.}

\item{remove_clusters}{Character vector. Cluster(s) to remove.}

\item{start_date}{Starting date of the study. If \code{NULL} (the default),
the date of the antares study is used.}

\item{opts}{List of simulation parameters returned by the function
\code{setSimulationPath}}
}
\value{
a \code{data.table}.
}
\description{
Get sum of PCN for stopped clusters by weeks
}
\examples{
\dontrun{

library(antaresRead)
setSimulationPath(path = "PATH/TO/STUDY")

pcn_stop_clusters()

pcn_stop_clusters(first_weekday = 6) # saturday

}
}
