% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_study}
\alias{setup_study}
\title{Create an (empty) Antares study for Extended Shutdown Simulation}
\usage{
setup_study(path, study_name = "prolongation-arrets",
  area_name = "area")
}
\arguments{
\item{path}{Path to a directory where to create study.}

\item{study_name}{Name of the study.}

\item{area_name}{Name of the area to create}
}
\description{
Create an (empty) Antares study for Extended Shutdown Simulation
}
